pyproject.toml
[project]
name = "bible"
version = "0.1.0"
description = ""
authors = [
    { name = "kiortir", email = "kiortir@yandex.ru" }
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "uvicorn[standart] (>=0.34.0,<0.35.0)",
    "fastapi[standart] (>=0.115.11,<0.116.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "sqlalchemy[asyncio] (>=2.0.39,<3.0.0)",
    "faststream[redis] (>=0.5.37,<0.6.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "redis[hiredis] (>=5.2.1,<6.0.0)",
    "yarl (>=1.18.3,<2.0.0)",
    "httpx-limiter (>=0.1.0,<0.2.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "qsparser (>=1.1.0,<2.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"

